It should not crash anymore

+++ src/patch_tools.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <dlfcn.h>
 
 
@@ -6,7 +7,7 @@
 //Return 0 on success, return 1 on failure.
 //4 HEX ie FF FF FF FF
 inline int PatchHex_32(void* baseaddress, int offset, uint32_t original_hex, uint32_t new_hex){
-    uint32_t* addrtocheck = (uint32_t*)baseaddress + offset;
+    uint32_t* addrtocheck = (uint32_t*)baseaddress + (long)offset;
     if (*addrtocheck == original_hex){
         *addrtocheck = new_hex;
         return 0;
